+++ CHANGELOG.md
@@ -7,6 +7,12 @@ Notable changes to the project will be documented here
 
 ### Changed
 
+## 0.0.6
+### Added
+- Adding Editor Context Menu Items to help setup custom worlds scripts:
+    - Assets/Custom World/Setup to setup all the required base files (Bootstrap, Enum and Attribute)
+    - Assets/Custom World/New Custom World to create a new custom world from one of the types in Enum
+
 ## 2020-09-04
 ### Removed
 - Removed Prototype folder from root folder